nosplit_example/requirements.txt
appnope==0.1.3
asttokens==2.4.1
backports.functools-lru-cache==1.6.5
Bottleneck==1.3.5
comm==0.1.4
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
django==4.1.7
entrypoints==0.4
exceptiongroup==1.1.3
executing==2.0.1
fonttools==4.25.0
importlib-resources==6.1.0
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
joblib==1.2.0
jupyter-client==7.3.4
jupyter_core==5.5.0
kiwisolver==1.4.4
matplotlib==3.8.0
matplotlib-inline==0.1.6
mysql-connector-python==8.1.0
mysqlclient==2.0.3
munkres==1.1.4
nest-asyncio==1.5.8
networkx==3.1
numexpr==2.8.7
numpy==1.26.0
packaging==23.2
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pip==23.3
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.0
scikit-learn==1.3.0
scipy==1.11.3
setuptools==68.0.0
six==1.16.0
stack-data==0.6.2
threadpoolctl==2.2.0
tornado==6.1
tqdm==4.65.0
traitlets==5.13.0
typing_extensions==4.8.0
tzdata==2023.3
wcwidth==0.2.9
wheel==0.41.2
zipp==3.11.0
